README Revert "[http:CMakeLists] Build | Required modules: `utility`"
This reverts commit d33c06d918f1bbbb45dfac9aa83f632741bb1ee6.

This commit reverts an implementation of the functionality that enables acquiring autopiot version using MAVLink. The implementation is correct, however, bugs in the communication scheme do not allow using it correctly. Revert this commit as soon as the comm. is fixed.

Revert "[http:version] Enh | AP versions acquision using the new module framework"

This reverts commit 60947b2bf7c13aeebd90ba1660171043a607100a.