[module:SdMemoryProvider] Bug | cJSON, entry replacement, type match
The previous implementation relied on the presupposition that calling "set
value" kind of methods upon cJSON objects would replace a current value w/ the
next one, regardless of its type. As it turns out, the type matters, and in the
case when the types don't match, the call will be ignored.